NumericLiteralCaseFixer - Introduction
