+++ doc/rules/casing/integer_literal_case.rst
@@ -4,6 +4,14 @@ Rule ``integer_literal_case``
 
 Integer literals must be in correct case.
 
+Warning
+-------
+
+This rule is deprecated and will be removed in the next major version
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+You should use ``numeric_literal_case`` instead.
+
 Examples
 --------
 
@@ -19,12 +27,3 @@ Example #1
    -$bar = 0B11111111;
    +$foo = 0xFF;
    +$bar = 0b11111111;
-
-Rule sets
----------
-
-The rule is part of the following rule sets:
-
-- `@PhpCsFixer <./../../ruleSets/PhpCsFixer.rst>`_
-- `@Symfony <./../../ruleSets/Symfony.rst>`_
-
