+++ tests/Fixer/Casing/IntegerLiteralCaseFixerTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 
 namespace PhpCsFixer\Tests\Fixer\Casing;
 
+use PhpCsFixer\Fixer\DeprecatedFixerInterface;
 use PhpCsFixer\Tests\Test\AbstractFixerTestCase;
 
 /**
@@ -23,86 +24,14 @@ use PhpCsFixer\Tests\Test\AbstractFixerTestCase;
  */
 final class IntegerLiteralCaseFixerTest extends AbstractFixerTestCase
 {
-    /**
-     * @dataProvider provideFixCases
-     */
-    public function testFix(string $expected, string $input = null): void
+    public function testFunctionIsDeprecatedProperly(): void
     {
-        $this->doTest($expected, $input);
-    }
+        $fixer = $this->fixer;
 
-    public static function provideFixCases(): iterable
-    {
-        yield [
-            '<?php $foo1 = 0xFF; $foo2 = 0xDEFA; $foo3 = 0xFA; $foo4 = 0xFA;',
-            '<?php $foo1 = 0XFF; $foo2 = 0xdefa; $foo3 = 0Xfa; $foo4 = 0xFA;',
-        ];
-
-        yield [
-            '<?php $foo = 0xA1FB20;',
-            '<?php $foo = 0xa1fb20;',
-        ];
-
-        yield [
-            '<?php $foo = -0xA1FB20;',
-            '<?php $foo = -0xa1fb20;',
-        ];
-
-        yield [
-            '<?php $foo = 0b1101;',
-            '<?php $foo = 0B1101;',
-        ];
-
-        yield [
-            '<?php $A = 1_234_567;',
-        ];
-
-        yield [
-            '<?php $A = +0xAA_FF_00;',
-            '<?php $A = +0Xaa_ff_00;',
-        ];
-
-        yield [
-            '<?php $A = -0x00_AA_FF_00;',
-            '<?php $A = -0X00_aa_ff_00;',
-        ];
-
-        yield 'bin_PHP_INT_MAX' => [
-            '<?php $foo = 0b111111111111111111111111111111111111111111111111111111111111111;',
-            '<?php $foo = 0B111111111111111111111111111111111111111111111111111111111111111;',
-        ];
-
-        yield 'hex_plus_PHP_INT_MAX' => [
-            '<?php $foo = +0x7FFFFFFFFFFFFFFF;',
-            '<?php $foo = +0X7fffffffffffffff;',
-        ];
-
-        yield 'hex_minus_PHP_INT_MAX' => [
-            '<?php $foo = -0x7FFFFFFFFFFFFFFF;',
-            '<?php $foo = -0X7fffffffffffffff;',
-        ];
-    }
-
-    /**
-     * @dataProvider provideFix81Cases
-     *
-     * @requires PHP 8.1
-     */
-    public function testFix81(string $expected, ?string $input = null): void
-    {
-        $this->doTest($expected, $input);
-    }
-
-    public static function provideFix81Cases(): iterable
-    {
-        yield [
-            '<?php $foo = 0o123;',
-            '<?php $foo = 0O123;',
-        ];
-
-        yield [
-            '<?php $foo = -0o123;',
-            '<?php $foo = -0O123;',
-        ];
+        self::assertInstanceOf(DeprecatedFixerInterface::class, $fixer);
+        self::assertSame(
+            ['numeric_literal_case'],
+            $fixer->getSuccessorsNames(),
+        );
     }
 }
